Finish graph

+++ leettrader/stock/routes.py
@@ -73,6 +73,12 @@ def search_page(code):
 @login_required
 def get_csv(code):
   tmp_bin = "leettrader/stock/tmp/"
+  '''
+  caching the csv file if called repeatedly
+  '''
+  if os.path.isfile("leettrader/stock/tmp/"+ code + ".csv"):
+    print("HERE")
+    return send_file("stock/tmp/"+ code + ".csv")
   for root, dirs, files in os.walk(tmp_bin):
       for f in files:
           os.remove(os.path.join(root, f))
